fix: wzj made

+++ cloudfunctions/delete_item/index.js
@@ -9,9 +9,12 @@ const db = cloud.database()
 
 exports.main = async (event, context) => {
   const wxContext = cloud.getWXContext()
+  await db.collection('photos').doc(event._id).remove()
   //从存储上删除文件
-  const result = await cloud.deleteFile({
-    fileList: [event.fileID],
-  })
-  return await db.collection('photos').doc(event._id).remove()
+  if (event.typpe == 'delete') {
+    return await cloud.deleteFile({
+      fileList: [event.fileID],
+    })
+  } else {
+  }
 }
